+++ CHANGELOG.md
@@ -9,3 +9,9 @@ Screen View, log event
 Screen View Flow, current screen
 ## 1.0.7
 Update input events, screens, events as json string
+## 1.0.8
+Fix bugs input event when setState, add referral id
+## 1.0.9
+Fix bugs input event when setState, add referral id
+## 1.1.0
+New method run speed test for ios